refactor(rvr-algebra): use common pad_modulus
